refactor: extract prompt metadata snapshots
